+++ src/media/rendering/erChannelRenderer.h
@@ -19,8 +19,8 @@ public:
     void setTestVideoPlayer(erSyncedVideoPlayer* _testVideoPlayer);
     void setVideoPlayers(map<string, ofPtr<erSyncedVideoPlayer>>* _videoPlayers);
 
-    void newOpeningGlitchPeriod(unsigned long long from, float duration);
-    void newClosingGlitchPeriod(unsigned long long from, float duration);
+    void newOpeningGlitchPeriod(u_int64_t from, float duration);
+    void newClosingGlitchPeriod(u_int64_t from, float duration);
     bool isChannelPlaying(int channel);
     void assign(int channel, erPlayParams params);
